+++ service/controller/resource/tccpnlifecycle/delete.go
@@ -1,0 +1,9 @@
+package tccpnlifecycle
+
+import (
+	"context"
+)
+
+func (r *Resource) EnsureDeleted(ctx context.Context, obj interface{}) error {
+	return nil
+}