+++ service/controller/resource/tccpnlifecycle/error.go
@@ -1,0 +1,21 @@
+package tccpnlifecycle
+
+import "github.com/giantswarm/microerror"
+
+var invalidConfigError = &microerror.Error{
+	Kind: "invalidConfigError",
+}
+
+// IsInsserts invalidConfigError.
+func IsInvalidConfig(err error) bool {
+	return microerror.Cause(err) == invalidConfigError
+}
+
+var invalidProviderIDError = &microerror.Error{
+	Kind: "invalidProviderID",
+}
+
+// IsInvalidProviderID asserts invalidConfigError.
+func IsInvalidProviderID(err error) bool {
+	return microerror.Cause(err) == invalidProviderIDError
+}