+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Complete master's EC2 lifecycle hook on upgrades.
+
 ## [14.7.1] - 2023-02-03
 
 - Switch container registry in China